separate cli.go in gcli style
